feat: split dal and bl within diracx-db
