+++ diracx-db/src/diracx/dal/exceptions.py
