+++ diracx-db/src/diracx/dal/sql/utils/base.py
