+++ diracx-db/src/diracx/dal/sql/utils/functions.py
